@@ -1,13 +1,13 @@
 package com.aihui.lib.base.api.retrofit.download;
 
 import android.os.SystemClock;
-import android.support.annotation.NonNull;
 
 import com.aihui.lib.base.util.HandlerUtils;
 import com.aihui.lib.base.util.LogUtils;
 
 import java.io.IOException;
 
+import androidx.annotation.NonNull;
 import okhttp3.MediaType;
 import okhttp3.RequestBody;
 import okio.Buffer;
