@@ -3,8 +3,6 @@ package com.aihui.lib.base.ui.view;
 import android.content.Context;
 import android.content.res.TypedArray;
 import android.graphics.Rect;
-import android.support.annotation.NonNull;
-import android.support.v4.widget.ViewDragHelper;
 import android.util.AttributeSet;
 import android.view.MotionEvent;
 import android.view.View;
@@ -12,10 +10,11 @@ import android.widget.RelativeLayout;
 
 import com.aihui.lib.base.R;
 
-import java.util.ArrayList;
 import java.util.HashMap;
-import java.util.List;
 import java.util.Map;
+
+import androidx.annotation.NonNull;
+import androidx.customview.widget.ViewDragHelper;
 
 /**
  * Created by zhy on 15/6/3.
