@@ -1,7 +1,8 @@
 package com.aihui.lib.base.ui.view.treeview;
 
-import android.support.v7.widget.RecyclerView;
 import android.view.View;
+
+import androidx.recyclerview.widget.RecyclerView;
 
 
 public abstract class TreeViewBinder<VH extends RecyclerView.ViewHolder> implements LayoutItemType {
