@@ -3,10 +3,11 @@ package com.aihui.lib.base.util;
 import android.app.NotificationChannel;
 import android.app.NotificationManager;
 import android.content.Context;
-import android.support.v4.app.NotificationCompat;
 
 import com.aihui.lib.base.R;
 import com.aihui.lib.base.app.BaseApplication;
+
+import androidx.core.app.NotificationCompat;
 
 /**
  * Created by 胡一鸣 on 2018/8/15.
