@@ -2,12 +2,13 @@ package com.aihui.lib.base.util;
 
 import android.os.Parcel;
 import android.os.Parcelable;
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
 
 import com.aihui.lib.base.app.BaseApplication;
 
 import java.io.File;
+
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
 
 public final class ParcelableUtils {
 
