@@ -2,13 +2,14 @@ package com.aihui.lib.nurse.ui.view.sort;
 
 import android.content.Context;
 import android.content.res.TypedArray;
-import android.support.annotation.Nullable;
 import android.util.AttributeSet;
 import android.view.View;
 import android.widget.ImageView;
 import android.widget.LinearLayout;
 
 import com.aihui.lib.nurse.R;
+
+import androidx.annotation.Nullable;
 
 /**
  * Created by 胡一鸣 on 2018/8/18.
