@@ -2,8 +2,6 @@ package com.aihui.lib.nurse.util;
 
 import android.content.ComponentCallbacks;
 import android.os.Parcelable;
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
 
 import com.aihui.lib.base.api.retrofit.BaseObserver;
 import com.aihui.lib.base.api.retrofit.RetrofitManager;
@@ -19,6 +17,8 @@ import org.reactivestreams.Publisher;
 import java.io.File;
 import java.util.List;
 
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
 import io.reactivex.Flowable;
 import io.reactivex.Observable;
 import io.reactivex.ObservableSource;
